Added ``IntrinsicFunction`` in ASR (#1616)

+++ integration_tests/test_builtin_abs.py
@@ -32,8 +32,8 @@ def test_abs():
 
     b: bool
     b = True
-    assert abs(b) == 1
+    assert abs(i32(b)) == 1
     b = False
-    assert abs(b) == 0
+    assert abs(i32(b)) == 0
 
 test_abs()
